# src/main/java/com/lottoland/api/JankenApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.43).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.lottoland.api;

import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.lottoland.dto.ErrorDTO;
import com.lottoland.dto.JankenDTO;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-07-07T22:14:06.712861+02:00[Europe/Madrid]")
@Validated
@CrossOrigin
public interface JankenApi {

    @Operation(summary = "Reset the session of janken for one user.", description = "", tags={ "Janken" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "text/plain", schema = @Schema(implementation = String.class))),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDTO.class))) })
    @RequestMapping(value = "/janken/reset/{sessionID}",
        produces = { "text/plain", "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<String> postResetJankenBySession(@Parameter(in = ParameterIn.PATH, description = "session generated by front simulating a user logged", required=true, schema=@Schema()) @PathVariable("sessionID") String sessionID);


    @Operation(summary = "Play a round of janken for one user.", description = "", tags={ "Janken" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation = JankenDTO.class))),
        
        @ApiResponse(responseCode = "400", description = "Invalid ID supplied", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDTO.class))) })
    @RequestMapping(value = "/janken/{sessionID}",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<JankenDTO> postRoundJankenBySession(@Parameter(in = ParameterIn.PATH, description = "session generated by front simulating a user logged", required=true, schema=@Schema()) @PathVariable("sessionID") String sessionID);

}