# src/main/java/com/lottoland/api/JankenApiController.java
package com.lottoland.api;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

import com.lottoland.dto.JankenDTO;
import com.lottoland.service.JankenService;

import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Schema;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-07-07T17:40:06.921194+02:00[Europe/Madrid]")
@RestController
public class JankenApiController implements JankenApi {

	@Autowired
	private JankenService jankenService;

    @org.springframework.beans.factory.annotation.Autowired
    public JankenApiController() {
    }

    public ResponseEntity<String> postResetJankenBySession(@Parameter(in = ParameterIn.PATH, description = "session generated by front simulating a user logged", required=true, schema=@Schema()) @PathVariable("sessionID") String sessionID) {
    	return new ResponseEntity<String>(jankenService.reset(sessionID) ,HttpStatus.OK);
    }

    public ResponseEntity<JankenDTO> postRoundJankenBySession(@Parameter(in = ParameterIn.PATH, description = "session generated by front simulating a user logged", required=true, schema=@Schema()) @PathVariable("sessionID") String sessionID) {
        return new ResponseEntity<JankenDTO>(jankenService.round(sessionID) ,HttpStatus.OK);
    }

}